fix: use dist.js when node > 6
