+++ package.json
@@ -2,11 +2,11 @@
   "name": "webpack-cdn-plugin",
   "version": "1.3.4",
   "description": "A webpack plugin that use externals of CDN urls for production and local node_modules for development",
-  "main": "dist.js",
+  "main": "index.js",
   "scripts": {
     "start": "webpack --config example/webpack.config.js",
     "test": "istanbul cover jasmine",
-    "build": "buble -i index.js -o dist.js",
+    "build": "buble -i src.js -o dist.js",
     "prepublish": "npm run build"
   },
   "repository": {
@@ -56,6 +56,7 @@
     "istanbul": "^0.4.5",
     "jasmine": "^2.6.0",
     "jasmine-spec-reporter": "^4.1.1",
+    "semver": "^5.3.0",
     "webpack": "^3.0.0"
   }
 }
